+++ .env.example
@@ -4,3 +4,4 @@ DATABASE_USER=
 DATABASE_PASSWORD=
 DATABASE_NAME=
 DATABASE_URL=
+EMAIL_PASSWORD_APP=
