+++ app/src/main/java/com/mealmatch/controller/TelaLoginController.java
@@ -4,9 +4,6 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-import org.apache.commons.mail.DefaultAuthenticator;
-import org.apache.commons.mail.SimpleEmail;
-
 import com.mealmatch.utils.EmailSender;
 
 import javafx.application.Platform;
